Refactor: 상수클래스 이름 변경 초기 생각과 달라 상수클래스명 변경
