+++ src/main/java/racingcar/domain/CarName.java
@@ -2,8 +2,8 @@ package racingcar.domain;
 
 import racingcar.exception.InputCarNameIllegarArgumentException;
 
-import static racingcar.utils.OptionConstants.CAR_NAME_MAX_LENGTH;
-import static racingcar.utils.OptionConstants.MSG_CAR_NAME_LENGTH_EXCEPTION;
+import static racingcar.utils.RacingCarConstants.CAR_NAME_MAX_LENGTH;
+import static racingcar.utils.RacingCarConstants.MSG_CAR_NAME_LENGTH_EXCEPTION;
 
 public class CarName {
     private final String carName;
