+++ src/test/java/racingcar/domain/CarNameTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static racingcar.utils.OptionConstants.MSG_CAR_NAME_LENGTH_EXCEPTION;
+import static racingcar.utils.RacingCarConstants.MSG_CAR_NAME_LENGTH_EXCEPTION;
 
 public class CarNameTest {
     @DisplayName("자동차이름 원시값 객체 생성을 통해 원시값이 포장되는지 확인")
